feat: add PlayerPanel, SocialPanel, FormPanel, GiftPanel, MessagePanel, and supporting components
- Implemented PlayerPanel for displaying player stats and equipment.
- Created SocialPanel for managing social interactions and displaying friend details.
- Developed FormPanel for handling form submissions with validation.
- Introduced GiftPanel for sending gifts with inventory management.
- Added MessagePanel for real-time messaging functionality.
- Established PanelFrame for consistent panel layout and styling.
- Created reusable components like GoldRow and InfoCard for better UI consistency.
- Defined styles for panels and components to enhance visual appearance.
