+++ app/globals.css
@@ -119,7 +119,8 @@ body {
   background: linear-gradient(
     105deg,
     transparent 30%,
-    rgba(255, 255, 255, 0.07) 50%,
+    rgba(120, 200, 255, 0.09) 48%,
+    rgba(200, 230, 255, 0.06) 52%,
     transparent 70%
   );
   pointer-events: none;
