+++ entities/nav/index.ts
@@ -1,2 +1,3 @@
 export * from "./types";
 export * from "./config";
+export * from "./generators";
