+++ features/lifelog/data.ts
@@ -1,0 +1,38 @@
+export const COLLECTION_DATA = [
+  { title: "Kirito 1/7 Scale Figure",         cat: "Figure",       qty: 1,  cond: "Mint in box",  from: "Good Smile" },
+  { title: "Asuna ALO Ver. Figure",            cat: "Figure",       qty: 1,  cond: "Mint",         from: "Kotobukiya" },
+  { title: "SAO Progressive Manga",            cat: "Manga",        qty: 8,  cond: "Like new",     from: "Kyobo" },
+  { title: "SAO Light Novel JP Complete",      cat: "Light Novel",  qty: 27, cond: "JP edition",   from: "Amazon JP" },
+  { title: "SAO Artbook Vol. 1",               cat: "Artbook",      qty: 1,  cond: "Like new",     from: "Comiket" },
+  { title: "Elucidator Replica Sword",         cat: "Merchandise",  qty: 1,  cond: "Display",      from: "Animax Shop" },
+  { title: "SAO Complete Series Bluray",       cat: "Bluray",       qty: 1,  cond: "Sealed",       from: "Aniplex" },
+  { title: "SAO Enamel Pin Set",               cat: "Merchandise",  qty: 12, cond: "Complete set", from: "CR Shop" },
+];
+
+export const MEDIA_DATA = [
+  { title: "Sword Art Online",                  cat: "Anime",  cur: 25,  total: 25,  status: "COMPLETED",    rating: 9.5 },
+  { title: "Attack on Titan",                   cat: "Anime",  cur: 87,  total: 87,  status: "COMPLETED",    rating: 9.8 },
+  { title: "Demon Slayer",                      cat: "Anime",  cur: 26,  total: 26,  status: "COMPLETED",    rating: 9.2 },
+  { title: "Clean Code",                        cat: "Book",   cur: 464, total: 464, status: "COMPLETED",    rating: 9.0 },
+  { title: "Fullmetal Alchemist: Brotherhood",  cat: "Anime",  cur: 48,  total: 64,  status: "WATCHING",     rating: null },
+  { title: "The Pragmatic Programmer",          cat: "Book",   cur: 100, total: 352, status: "READING",      rating: null },
+  { title: "Your Name",                         cat: "Movie",  cur: 1,   total: 1,   status: "COMPLETED",    rating: 10.0 },
+  { title: "Overlord",                          cat: "Anime",  cur: 0,   total: 13,  status: "PLAN_TO_WATCH",rating: null },
+  { title: "Designing Data-Intensive Apps",     cat: "Book",   cur: 250, total: 616, status: "READING",      rating: null },
+  { title: "Spirited Away",                     cat: "Movie",  cur: 1,   total: 1,   status: "COMPLETED",    rating: 9.9 },
+];
+
+export const EXERCISE_DATA = [
+  { cat: "Running",          dur: 32,  dist: 5.2,  cal: 310, date: "2026-03-02" },
+  { cat: "Cycling",          dur: 75,  dist: 28.4, cal: 620, date: "2026-03-01" },
+  { cat: "Strength Training",dur: 55,  dist: null, cal: 380, date: "2026-02-28" },
+  { cat: "Swimming",         dur: 40,  dist: 1.5,  cal: 420, date: "2026-02-26" },
+  { cat: "Running",          dur: 28,  dist: 4.8,  cal: 285, date: "2026-02-25" },
+  { cat: "Yoga",             dur: 45,  dist: null, cal: 120, date: "2026-02-24" },
+  { cat: "Hiking",           dur: 180, dist: 12.0, cal: 850, date: "2026-02-22" },
+  { cat: "Strength Training",dur: 60,  dist: null, cal: 420, date: "2026-02-21" },
+  { cat: "Running",          dur: 35,  dist: 6.0,  cal: 360, date: "2026-02-19" },
+  { cat: "Cycling",          dur: 90,  dist: 35.2, cal: 750, date: "2026-02-16" },
+  { cat: "Jump Rope",        dur: 20,  dist: null, cal: 200, date: "2026-02-15" },
+  { cat: "Swimming",         dur: 50,  dist: 2.0,  cal: 530, date: "2026-02-12" },
+];
