+++ features/player/data.ts
@@ -1,0 +1,65 @@
+export const ACHIEVEMENT_DATA = [
+  { code: "FIRST_BLOOD",   name: "First Blood",           cat: "Combat",      desc: "Defeated your first monster." },
+  { code: "FLOOR_BOSS",    name: "Floor Boss Slayer",      cat: "Combat",      desc: "Defeated a floor boss." },
+  { code: "SOLO_RUN",      name: "Solo Runner",            cat: "Exploration", desc: "Cleared a dungeon entirely solo." },
+  { code: "SPEED_RUN",     name: "Speedrunner",            cat: "Completion",  desc: "Cleared a floor in record time." },
+  { code: "BLACKSMITH_1",  name: "Blacksmith Apprentice",  cat: "Crafting",    desc: "Crafted your first item." },
+  { code: "DUAL_WIELD",    name: "Dual Wield Mastery",     cat: "Combat",      desc: "Unlocked the Dual Wield skill." },
+  { code: "PERFECT_GUARD", name: "Perfect Guard",          cat: "Defense",     desc: "Blocked 100 attacks with perfect timing." },
+  { code: "MONSTER_HUNT",  name: "Monster Hunter",         cat: "Combat",      desc: "Defeated 1000 monsters." },
+  { code: "TREASURE_HUNT", name: "Treasure Hunter",        cat: "Exploration", desc: "Found 50 hidden treasure rooms." },
+  { code: "BEATER",        name: "The Beater",             cat: "Special",     desc: "Recognized as a beta tester." },
+  { code: "GUILD_FOUND",   name: "Guild Founder",          cat: "Social",      desc: "Founded your first guild." },
+  { code: "MKT_MASTER",    name: "Market Master",          cat: "Economy",     desc: "Completed 100 market trades." },
+  { code: "LVL_50",        name: "Level 50 Milestone",     cat: "Growth",      desc: "Reached character level 50." },
+  { code: "ENCHANT",       name: "Enchantment Master",     cat: "Crafting",    desc: "Enchanted 50 items successfully." },
+  { code: "NIGHT_RUN",     name: "Night Runner",           cat: "Exploration", desc: "Explored 10 dungeons at night." },
+  { code: "PARTY_LEAD",    name: "Party Leader",           cat: "Social",      desc: "Led a party to a boss clear." },
+  { code: "RARE_ITEM",     name: "Rare Collector",         cat: "Collection",  desc: "Obtained 20 rare+ items." },
+  { code: "LEGEND_KILL",   name: "Legend Slayer",          cat: "Combat",      desc: "Defeated a legendary-tier monster." },
+];
+
+export const CERTIFICATION_DATA = [
+  { name: "AWS Solutions Architect",    issuer: "Amazon Web Services",       cat: "Cloud",       expires: "2028-06-15" },
+  { name: "Python Professional",        issuer: "Python Institute",           cat: "Programming", expires: null },
+  { name: "Kubernetes Administrator",   issuer: "CNCF",                       cat: "DevOps",      expires: "2027-09-10" },
+  { name: "TOEIC 935",                  issuer: "ETS Korea",                  cat: "Language",    expires: "2026-11-02" },
+  { name: "Google Cloud Professional",  issuer: "Google",                     cat: "Cloud",       expires: "2027-12-01" },
+  { name: "React Developer Certification", issuer: "Meta",                    cat: "Frontend",    expires: null },
+  { name: "Computer Science B.S.",      issuer: "Seoul National University",  cat: "Academic",    expires: null },
+  { name: "SQLD Database Developer",    issuer: "Korea Data Agency",          cat: "Database",    expires: null },
+  { name: "Security+ Certified",        issuer: "CompTIA",                    cat: "Security",    expires: "2028-10-05" },
+  { name: "TypeScript Advanced",        issuer: "Microsoft",                  cat: "Programming", expires: null },
+  { name: "Docker Certified Associate", issuer: "Docker Inc.",                cat: "DevOps",      expires: "2027-05-28" },
+  { name: "JLPT N3 Japanese",           issuer: "Japan Foundation",           cat: "Language",    expires: null },
+];
+
+export const TITLE_DATA = [
+  { code: "BLACK_SWORDSMAN", name: "Black Swordsman", cat: "Achievement", desc: "The legendary solo player." },
+  { code: "BEATER",          name: "Beater",           cat: "Special",     desc: "A beta tester with prior knowledge." },
+  { code: "SOLO_KING",       name: "Solo King",        cat: "Exploration", desc: "Cleared 10 floors solo." },
+  { code: "FLOOR_CLEARER",   name: "Floor Clearer",    cat: "Combat",      desc: "Front-line floor clearing." },
+  { code: "DUAL_WIELDER",    name: "Dual Wielder",     cat: "Combat",      desc: "Holder of Dual Wield skill." },
+  { code: "MASTER_CRAFTER",  name: "Master Crafter",   cat: "Crafting",    desc: "Crafted masterwork items." },
+  { code: "GUILD_CHIEF",     name: "Guild Chief",      cat: "Social",      desc: "Leader of a recognized guild." },
+  { code: "LEGEND_SLAYER",   name: "Legend Slayer",    cat: "Combat",      desc: "Slayer of legendary monsters." },
+  { code: "MARKET_KING",     name: "Market King",      cat: "Economy",     desc: "100 trades, perfect ratings." },
+  { code: "NIGHT_WALKER",    name: "Night Walker",     cat: "Exploration", desc: "Nocturnal dungeon master." },
+  { code: "FRONTLINER",      name: "Frontliner",       cat: "Rank",        desc: "Active front-line team member." },
+  { code: "RARE_HUNTER",     name: "Rare Hunter",      cat: "Collection",  desc: "Collector of rare items." },
+];
+
+export const HOBBY_DATA = [
+  { name: "Programming",      cat: "Tech",          custom: "Full-Stack Dev",    proficiency: 85, status: "ACTIVE",   xp: 42000 },
+  { name: "Reading",          cat: "Learning",      custom: "Tech & Fiction",    proficiency: 72, status: "ACTIVE",   xp: 28500 },
+  { name: "Gaming",           cat: "Entertainment", custom: "VR & JRPG",         proficiency: 90, status: "ACTIVE",   xp: 65000 },
+  { name: "Running",          cat: "Fitness",       custom: "Morning Runs",      proficiency: 60, status: "ACTIVE",   xp: 12000 },
+  { name: "Cooking",          cat: "Lifestyle",     custom: "Japanese Cuisine",  proficiency: 55, status: "ACTIVE",   xp: 9800 },
+  { name: "Drawing",          cat: "Art",           custom: "Anime Sketching",   proficiency: 40, status: "ON_HOLD",  xp: 5200 },
+  { name: "Guitar",           cat: "Music",         custom: "Acoustic Guitar",   proficiency: 48, status: "ON_HOLD",  xp: 7400 },
+  { name: "Photography",      cat: "Art",           custom: "Street & Nature",   proficiency: 63, status: "ACTIVE",   xp: 11200 },
+  { name: "Language Learning",cat: "Learning",      custom: "Japanese N3→N2",   proficiency: 65, status: "ACTIVE",   xp: 18600 },
+  { name: "Cycling",          cat: "Fitness",       custom: "Road Cycling",      proficiency: 58, status: "ACTIVE",   xp: 8900 },
+  { name: "Meditation",       cat: "Wellness",      custom: "Mindfulness",       proficiency: 45, status: "ACTIVE",   xp: 4300 },
+  { name: "Origami",          cat: "Art",           custom: "Complex Origami",   proficiency: 35, status: "INACTIVE", xp: 2100 },
+];
