+++ shared/api/types/auth.ts
@@ -1,0 +1,6 @@
+export interface AuthUser {
+  id: number;
+  email: string;
+  nickname: string;
+  role: "user" | "admin";
+}
