+++ shared/api/types/common.ts
@@ -1,0 +1,7 @@
+export interface ApiPage<T> {
+  content: T[];
+  page: number;
+  size: number;
+  totalElements: number;
+  totalPages: number;
+}
