+++ shared/design/icons.ts
@@ -1,0 +1,56 @@
+export const SAO_ICON = {
+  player:    '/icons/sao/Man.svg',
+  playerOn:  '/icons/sao/Man_on.svg',
+  skills:    '/icons/sao/Skills.svg',
+  skillsOn:  '/icons/sao/Skills_on.svg',
+  items:     '/icons/sao/Items.svg',
+  itemsOn:   '/icons/sao/Items_on.svg',
+  quest:     '/icons/sao/Quest & Message Box.svg',
+  questOn:   '/icons/sao/Quest & Message Box_on.svg',
+  social:    '/icons/sao/Men.svg',
+  socialOn:  '/icons/sao/Men_on.svg',
+  lifelog:   '/icons/sao/List.svg',
+  lifelogOn: '/icons/sao/List_on.svg',
+  market:    '/icons/sao/Details.svg',
+  marketOn:  '/icons/sao/Details_on.svg',
+  config:    '/icons/sao/Config.svg',
+  configOn:  '/icons/sao/Config_on.svg',
+  option:    '/icons/sao/Option.svg',
+  optionOn:  '/icons/sao/Option_on.svg',
+  logout:    '/icons/sao/Logout.svg',
+  logoutOn:  '/icons/sao/Logout_on.svg',
+  yes:       '/icons/sao/Yes.svg',
+  yesOn:     '/icons/sao/Yes_on.svg',
+  no:        '/icons/sao/No.svg',
+  noOn:      '/icons/sao/No_on.svg',
+  plus:      '/icons/sao/Plus.svg',
+  plusOn:    '/icons/sao/Plus_on.svg',
+  minus:     '/icons/sao/Minus.svg',
+  minusOn:   '/icons/sao/Minus_on.svg',
+  details:   '/icons/sao/Details.svg',
+  detailsOn: '/icons/sao/Details_on.svg',
+  list:      '/icons/sao/List.svg',
+  listOn:    '/icons/sao/List_on.svg',
+  start:     '/icons/sao/Start.svg',
+  startOn:   '/icons/sao/Start_on.svg',
+  calling:   '/icons/sao/Calling.svg',
+  callingOn: '/icons/sao/Calling_on.svg',
+  cancel:    '/icons/sao/Cancel & Dissolve.svg',
+  cancelOn:  '/icons/sao/Cancel & Dissolve_on.svg',
+  man:       '/icons/sao/Man.svg',
+  manOn:     '/icons/sao/Man_on.svg',
+  men:       '/icons/sao/Men.svg',
+  menOn:     '/icons/sao/Men_on.svg',
+  message:   '/icons/sao/Message.svg',
+  messageOn: '/icons/sao/Message_on.svg',
+  friend:    '/icons/sao/Friend.svg',
+  friendOn:  '/icons/sao/Friend_on.svg',
+  guild:     '/icons/sao/Guild.svg',
+  guildOn:   '/icons/sao/Guild_on.svg',
+  equipment: '/icons/sao/Equipment.svg',
+  equipOn:   '/icons/sao/Equipment_on.svg',
+  help:      '/icons/sao/Help & Unknown.svg',
+  helpOn:    '/icons/sao/Help & Unknown_on.svg',
+  invite:    '/icons/sao/Invite.svg',
+  inviteOn:  '/icons/sao/Invite_on.svg',
+} as const
