+++ shared/ui/PanelCard.tsx
@@ -215,7 +215,7 @@ export default function PanelCard({
         <div
           className="absolute inset-0 flex items-center justify-end"
           style={{
-            background: `rgba(224,62,99,${0.85 + swipeDeleteProgress * 0.15})`,
+            background: `rgba(244,63,94,${0.82 + swipeDeleteProgress * 0.18})`,
             paddingRight: 20,
           }}
           aria-hidden
@@ -275,26 +275,42 @@ export default function PanelCard({
           style={{
             height: UI_CONSTS.rightPanels.rowHeight,
             background: showActions
-              ? `rgba(255,198,68,0.14)`
+              ? `rgba(28,24,18,0.97)`
               : selected
-              ? "linear-gradient(180deg, rgba(255,198,68,0.97), rgba(240,178,44,0.96))"
-              : "linear-gradient(180deg, rgba(52,55,63,0.92), rgba(45,48,56,0.90))",
+              ? "linear-gradient(165deg, rgba(28,24,18,0.97), rgba(22,18,14,0.96))"
+              : "linear-gradient(165deg, rgba(16,16,18,0.96), rgba(12,12,14,0.94))",
             borderColor: showActions
-              ? `rgba(249,208,105,0.7)`
+              ? `rgba(225,195,85,0.70)`
               : isLongPressing
-              ? `rgba(249,208,105,0.5)`
+              ? `rgba(218,185,70,0.65)`
               : selected
-              ? "rgba(255,215,80,0.85)"
-              : "rgba(72,76,88,0.55)",
+              ? "rgba(218,178,55,0.58)"
+              : "rgba(185,185,192,0.20)",
             boxShadow: selected
-              ? `inset 0 0 0 1px rgba(255,222,100,0.4), 0 0 16px rgba(244,188,58,0.1)`
-              : "inset 0 0 0 1px rgba(255,255,255,0.04)",
-            transition: "border-color 0.6s ease, background 0.3s ease",
+              ? [
+                  "inset 0 0 0 1px rgba(218,178,60,0.14)",
+                  "0 0 20px rgba(218,178,55,0.10)",
+                ].join(", ")
+              : isLongPressing
+              ? `inset 0 0 0 1px rgba(218,185,70,0.18), 0 0 12px rgba(218,185,70,0.10)`
+              : `inset 0 0 0 1px rgba(255,255,255,0.04)`,
+            transition: "border-color 0.5s ease, background 0.3s ease, box-shadow 0.4s ease",
           }}
         >
           <div
             className="absolute inset-y-0 left-0 w-[5px]"
-            style={{ background: selected ? "rgba(244,192,72,0.95)" : "rgba(78,82,95,0.55)" }}
+            style={{
+              background: selected
+                ? "linear-gradient(180deg, rgba(255,232,120,0.98), rgba(228,168,28,0.96))"
+                : showActions
+                ? "rgba(248,197,78,0.65)"
+                : "rgba(155,157,165,0.38)",
+              boxShadow: selected
+                ? "0 0 10px rgba(244,188,48,0.55), inset 0 1px 0 rgba(255,248,180,0.60)"
+                : showActions
+                ? "0 0 6px rgba(248,197,78,0.35)"
+                : "none",
+            }}
           />
           <div style={{ paddingLeft: UI_CONSTS.rightPanels.cardPaddingX }}>
             <IconSlot label={slotLabel} active={selected} size={44} subtle />
@@ -309,7 +325,7 @@ export default function PanelCard({
             <p
               className="break-words font-semibold uppercase tracking-[0.08em]"
               style={{
-                color: selected ? "rgba(15,14,10,0.92)" : "rgba(200,206,218,0.90)",
+                color: selected ? "rgba(245,218,152,0.96)" : "rgba(218,220,228,0.92)",
                 fontSize: compact ? "0.9rem" : "1.1rem",
                 lineHeight: compact ? 1.2 : 1.15,
                 display: "-webkit-box",
@@ -324,7 +340,7 @@ export default function PanelCard({
               <p
                 className="mt-1 break-words text-sm tracking-[0.08em]"
                 style={{
-                  color: rarityColor(subtitle) ?? (selected ? "rgba(30,28,22,0.72)" : "rgba(130,140,158,0.80)"),
+                  color: rarityColor(subtitle) ?? (selected ? "rgba(188,162,100,0.82)" : "rgba(138,140,150,0.78)"),
                   display: "-webkit-box",
                   WebkitLineClamp: compact ? 1 : 2,
                   WebkitBoxOrient: "vertical",
@@ -356,14 +372,15 @@ export default function PanelCard({
               transition={{ type: "spring", stiffness: 400, damping: 28 }}
               className="absolute inset-0 z-[400001] flex items-center justify-end gap-2 overflow-hidden rounded-[2px] border"
               style={{
-                background: "linear-gradient(180deg, rgba(42,45,52,0.98), rgba(36,39,46,0.97))",
-                borderColor: `rgba(249,208,105,0.7)`,
+                background: "linear-gradient(180deg, rgba(14,14,16,0.98), rgba(10,10,12,0.97))",
+                borderColor: `rgba(220,185,70,0.55)`,
+                boxShadow: "inset 0 0 0 1px rgba(255,255,255,0.04), 0 0 20px rgba(220,185,70,0.06)",
                 paddingInline: 12,
               }}
             >
               <span
                 className="mr-auto uppercase"
-                style={{ fontSize: "10px", letterSpacing: "0.2em", color: "rgba(140,150,168,0.85)" }}
+                style={{ fontSize: "10px", letterSpacing: "0.2em", color: "rgba(148,150,158,0.70)" }}
               >
                 Action
               </span>
