+++ widgets/orb-nav/OrbNav.tsx
@@ -6,8 +6,8 @@ import { AnimatePresence, motion } from "framer-motion";
 import type { MainNavId } from "@/entities/nav";
 import { MOTION } from "@/shared/lib/motion";
 import { UI_CONSTS } from "@/shared/lib/uiConsts";
-
-import IconSlot from "@/shared/ui/IconSlot";
+import { SAO_ICON } from "@/shared/design/tokens";
+import { usePanScroll } from "@/shared/hooks/usePanScroll";
 
 type OrbItem = {
   id: MainNavId;
@@ -25,9 +25,35 @@ type OrbNavProps = {
 
 type Ripple = { id: number };
 
+// 기본 (비선택) 아이콘 — 흰 원 + 회색 픽토그램
+const ORB_ICONS: Partial<Record<MainNavId, string>> = {
+  player:    SAO_ICON.player,
+  skills:    SAO_ICON.skills,
+  inventory: SAO_ICON.items,
+  quests:    SAO_ICON.quest,
+  social:    SAO_ICON.social,
+  lifelog:   SAO_ICON.lifelog,
+  market:    SAO_ICON.market,
+  system:    SAO_ICON.config,
+};
+
+// 선택 상태 아이콘 — 금색 원 + 흰 픽토그램
+const ORB_ICONS_ON: Partial<Record<MainNavId, string>> = {
+  player:    SAO_ICON.playerOn,
+  skills:    SAO_ICON.skillsOn,
+  inventory: SAO_ICON.itemsOn,
+  quests:    SAO_ICON.questOn,
+  social:    SAO_ICON.socialOn,
+  lifelog:   SAO_ICON.lifelogOn,
+  market:    SAO_ICON.marketOn,
+  system:    SAO_ICON.configOn,
+};
+
 export default function OrbNav({ items, selectedId, onSelect, zIndex, onFocus }: OrbNavProps) {
   const scrollRef = useRef<HTMLDivElement>(null);
   const [ripples, setRipples] = useState<Ripple[]>([]);
+
+  usePanScroll(scrollRef);
 
   const {
     orbSize,
@@ -55,7 +81,6 @@ export default function OrbNav({ items, selectedId, onSelect, zIndex, onFocus }:
   const addRipple = () => {
     const id = Date.now();
     setRipples((prev) => [...prev, { id }]);
-    // 리플이 사라진 후 정리
     setTimeout(() => setRipples((prev) => prev.filter((r) => r.id !== id)), 650);
   };
 
@@ -66,61 +91,70 @@ export default function OrbNav({ items, selectedId, onSelect, zIndex, onFocus }:
       style={zIndex ? { zIndex } : undefined}
     >
       <div
-        className="relative flex items-center justify-center overflow-hidden rounded-full px-3"
+        className="relative flex items-center justify-center overflow-hidden px-3"
         style={{
           width: UI_CONSTS.layout.centerWidth,
           height: viewportHeight + framePaddingY * 2,
-          border: "1px solid rgba(82,127,214,0.22)",
-          background: "rgba(8,12,24,0.55)",
-          backdropFilter: "blur(10px)",
-          WebkitBackdropFilter: "blur(10px)",
+          borderRadius: "18px",
+          border: "1px solid rgba(200,200,205,0.22)",
+          background: "linear-gradient(155deg, rgba(18,18,20,0.97), rgba(14,14,16,0.95))",
+          backdropFilter: "blur(14px)",
+          WebkitBackdropFilter: "blur(14px)",
           boxShadow: [
-            "inset 0 0 0 1px rgba(120,160,255,0.08)",
-            "inset 0 0 60px rgba(0,0,0,0.40)",
-            "0 0 0 1px rgba(82,127,214,0.10)",
-            "0 0 44px rgba(82,127,214,0.14)",
+            "inset 0 0 0 1px rgba(255,255,255,0.04)",
+            "0 14px 40px rgba(0,0,0,0.65)",
           ].join(", "),
         }}
       >
+        {/* 시안 탑 액센트 */}
         <div
-          className="pointer-events-none absolute left-1/2 w-px -translate-x-1/2 bg-white/8"
-          style={{ top: framePaddingY, bottom: framePaddingY }}
+          aria-hidden
+          style={{
+            position: "absolute",
+            top: 0, left: "12%", right: "12%",
+            height: "1px",
+            background: "linear-gradient(90deg, transparent, rgba(218,178,60,0.55) 40%, rgba(218,178,60,0.55) 60%, transparent)",
+            pointerEvents: "none",
+          }}
         />
+        {/* 바텀 액센트 */}
+        <div
+          aria-hidden
+          style={{
+            position: "absolute",
+            bottom: 0, left: "12%", right: "12%",
+            height: "1px",
+            background: "linear-gradient(90deg, transparent, rgba(180,145,45,0.22) 40%, rgba(180,145,45,0.22) 60%, transparent)",
+            pointerEvents: "none",
+          }}
+        />
+
         <div
           ref={scrollRef}
           data-no-pan
           className="scrollbar-hide relative w-full overflow-y-auto overflow-x-hidden"
-          style={{
-            height: viewportHeight,
-            scrollbarWidth: "none",
-            msOverflowStyle: "none",
-          }}
+          style={{ height: viewportHeight, scrollbarWidth: "none", msOverflowStyle: "none" }}
         >
           <motion.div
             initial={{ opacity: 0.96 }}
             animate={{ opacity: 1 }}
             transition={MOTION.orbTrack.transition}
             className="flex flex-col items-center"
-            style={{
-              paddingTop: safePaddingY,
-              paddingBottom: safePaddingY,
-              gap: orbGap,
-            }}
+            style={{ paddingTop: safePaddingY, paddingBottom: safePaddingY, gap: orbGap }}
           >
             {items.map((item) => {
               const selected = item.id === selectedId;
+              const iconSrc    = ORB_ICONS[item.id];
+              const iconSrcOn  = ORB_ICONS_ON[item.id];
+              const activeSrc  = selected ? (iconSrcOn ?? iconSrc) : iconSrc;
 
               return (
                 <motion.button
                   layout="position"
                   key={item.id}
                   type="button"
-                  onClick={() => {
-                    addRipple();
-                    onSelect(item.id);
-                  }}
-                  // 선택 → 살짝 줄어드는 스냅, 호버 → 올라오는 느낌
-                  whileHover={{ scale: 1.07 }}
+                  onClick={() => { addRipple(); onSelect(item.id); }}
+                  whileHover={{ scale: 1.08 }}
                   whileTap={{ scale: 0.88 }}
                   transition={{
                     layout: { type: "tween", duration: 0.36, ease: [0.22, 1, 0.36, 1] },
@@ -131,60 +165,50 @@ export default function OrbNav({ items, selectedId, onSelect, zIndex, onFocus }:
                   aria-pressed={selected}
                   aria-label={item.label}
                 >
-                  {/* ── 선택 글로우 — framer-motion 키프레임으로 숨쉬기 ── */}
-                  <motion.div
-                    className="absolute rounded-full"
-                    animate={
-                      selected
-                        ? { opacity: [0.55, 1, 0.55], scale: [0.88, 1.14, 0.88] }
-                        : { opacity: 0, scale: 1 }
-                    }
-                    transition={
-                      selected
-                        ? {
-                            duration: 2.8,
-                            repeat: Infinity,
-                            ease: "easeInOut",
-                            repeatType: "loop",
-                          }
-                        : { duration: 0.28 }
-                    }
-                    style={{
-                      width: orbSize + 20,
-                      height: orbSize + 20,
-                      background:
-                        "radial-gradient(circle, rgba(248,201,87,0.30) 0%, rgba(248,201,87,0.12) 50%, rgba(248,201,87,0) 75%)",
-                      filter: "blur(2px)",
-                    }}
-                  />
-
-                  {/* ── 오브 컨테이너 (테두리 + 리플 + 아이콘) ────────── */}
+                  {/* 오브 컨테이너 — SVG가 원형 배경을 포함하므로 래퍼는 최소한 */}
                   <div
-                    className="relative rounded-full p-[4px]"
-                    style={{
-                      border: selected
-                        ? "1px solid rgba(249, 208, 105, 0.9)"
-                        : "1px solid rgba(229, 236, 246, 0.68)",
-                      boxShadow: selected
-                        ? "0 0 0 1px rgba(248,197,78,0.35), 0 0 22px rgba(247,196,70,0.20)"
-                        : "0 0 0 1px rgba(255,255,255,0.03)",
-                      background: selected
-                        ? "rgba(248,197,78,0.06)"
-                        : "rgba(0,0,0,0.22)",
-                      transition: "border 0.25s ease, box-shadow 0.25s ease, background 0.25s ease",
-                    }}
+                    className="relative"
+                    style={{ width: orbSize, height: orbSize, flexShrink: 0 }}
                   >
-                    {/* 클릭 리플 — overflow hidden으로 잘림 */}
+                    {/* 선택 글로우 — 배경에서 숨쉬는 금빛 광원 */}
+                    <motion.div
+                      className="pointer-events-none absolute rounded-full"
+                      animate={
+                        selected
+                          ? { opacity: [0.35, 0.75, 0.35], scale: [0.95, 1.25, 0.95] }
+                          : { opacity: 0, scale: 1 }
+                      }
+                      transition={
+                        selected
+                          ? { duration: 2.8, repeat: Infinity, ease: "easeInOut", repeatType: "loop" }
+                          : { duration: 0.28 }
+                      }
+                      style={{
+                        inset: -18,
+                        background: "radial-gradient(circle, rgba(248,180,40,0.40) 0%, rgba(248,180,40,0.14) 48%, rgba(248,180,40,0) 72%)",
+                        filter: "blur(4px)",
+                      }}
+                    />
+
+                    {/* 호버 시안 글로우 */}
                     <div
-                      className="pointer-events-none absolute inset-0 overflow-hidden rounded-full"
-                    >
+                      className="pointer-events-none absolute rounded-full opacity-0 transition-opacity duration-200 group-hover:opacity-100"
+                      style={{
+                        inset: -12,
+                        background: "radial-gradient(circle, rgba(220,200,140,0.18) 0%, rgba(200,178,90,0.06) 50%, transparent 72%)",
+                        filter: "blur(3px)",
+                      }}
+                    />
+
+                    {/* 리플 클리핑 레이어 */}
+                    <div className="pointer-events-none absolute inset-0 overflow-hidden rounded-full">
                       <AnimatePresence>
                         {ripples.map((r) => (
                           <motion.div
                             key={r.id}
                             className="absolute inset-0 rounded-full"
                             style={{ background: "rgba(248,197,78,0.40)" }}
-                            initial={{ scale: 0.3, opacity: 0.8 }}
+                            initial={{ scale: 0.3, opacity: 0.75 }}
                             animate={{ scale: 3.5, opacity: 0 }}
                             exit={{}}
                             transition={{ duration: 0.55, ease: "easeOut" }}
@@ -193,26 +217,87 @@ export default function OrbNav({ items, selectedId, onSelect, zIndex, onFocus }:
                       </AnimatePresence>
                     </div>
 
-                    <IconSlot
-                      label={item.slotLabel}
-                      active={selected}
-                      size={UI_CONSTS.orbNav.orbSize}
-                    />
+                    {/* 아이콘 이미지 — SVG에 원형 배경 내장, 필터 없이 원본 색상 사용 */}
+                    {activeSrc ? (
+                      <img
+                        src={activeSrc}
+                        alt={item.label}
+                        width={orbSize}
+                        height={orbSize}
+                        draggable={false}
+                        style={{
+                          display: "block",
+                          width: orbSize,
+                          height: orbSize,
+                          opacity: selected ? 1 : 0.72,
+                          transition: "opacity 0.28s ease",
+                          position: "relative",
+                          zIndex: 1,
+                        }}
+                        onError={(e) => {
+                          // SVG 로드 실패 시 텍스트 레이블로 대체
+                          const img = e.currentTarget;
+                          const span = document.createElement("span");
+                          span.textContent = item.slotLabel;
+                          span.style.cssText = [
+                            `width:${orbSize}px`,
+                            `height:${orbSize}px`,
+                            "display:grid",
+                            "place-items:center",
+                            "border-radius:9999px",
+                            `background:${selected ? "rgba(248,197,78,0.15)" : "rgba(0,40,90,0.55)"}`,
+                            `border:1px solid ${selected ? "rgba(249,208,105,0.70)" : "rgba(0,155,215,0.32)"}`,
+                            `font-size:${Math.max(10, Math.round(orbSize * 0.22))}px`,
+                            "font-weight:600",
+                            "letter-spacing:0.12em",
+                            `color:${selected ? "rgba(255,235,184,0.95)" : "rgba(200,220,245,0.80)"}`,
+                            "position:relative",
+                            "z-index:1",
+                          ].join(";");
+                          img.parentElement?.replaceChild(span, img);
+                        }}
+                      />
+                    ) : (
+                      // iconSrc가 없는 경우 텍스트 레이블
+                      <div
+                        className="grid place-items-center rounded-full"
+                        style={{
+                          width: orbSize,
+                          height: orbSize,
+                          background: selected
+                            ? "radial-gradient(circle, rgba(255,215,100,0.20), rgba(248,197,78,0.08) 60%, transparent)"
+                            : "radial-gradient(circle, rgba(0,70,150,0.30), rgba(0,30,80,0.14) 60%, transparent)",
+                          border: `1px solid ${selected ? "rgba(249,208,105,0.70)" : "rgba(0,155,215,0.32)"}`,
+                          position: "relative",
+                          zIndex: 1,
+                        }}
+                      >
+                        <span
+                          className="select-none font-semibold tracking-[0.14em]"
+                          style={{
+                            fontSize: Math.max(10, Math.round(orbSize * 0.24)),
+                            color: selected ? "rgba(255,235,184,0.95)" : "rgba(200,220,245,0.80)",
+                            lineHeight: 1,
+                          }}
+                        >
+                          {item.slotLabel}
+                        </span>
+                      </div>
+                    )}
                   </div>
 
-                  {/* ── 레이블 ──────────────────────────────────────────── */}
+                  {/* 레이블 */}
                   <motion.span
-                    animate={selected ? { opacity: 1 } : { opacity: 0.72 }}
-                    className="flex items-center justify-center text-center text-xs uppercase tracking-[0.28em]"
+                    animate={selected ? { opacity: 1 } : { opacity: 0.60 }}
+                    className="flex items-center justify-center text-center text-xs uppercase tracking-[0.26em]"
                     style={{
                       marginTop: labelGap,
                       minHeight: labelHeight + labelPaddingY * 2,
                       lineHeight: 1.2,
                       paddingInline: 4,
                       paddingBlock: labelPaddingY,
-                      color: selected
-                        ? "rgba(248, 220, 152, 0.95)"
-                        : "rgba(220, 228, 240, 0.72)",
+                      color: selected ? "rgba(248, 220, 152, 0.95)" : "rgba(175, 178, 188, 0.72)",
+                      textShadow: selected ? "0 0 10px rgba(248,200,100,0.35)" : "none",
                       transition: "color 0.25s ease",
                     }}
                   >
